+++ src/repositories/task_repository.py
@@ -4,16 +4,37 @@ from database_connection import get_database_connection
 from entities.task import Task
 
 class TaskRepository():
+    """Tehtäviin liittyvistä tietokantaoperaatioista vastaava luokka.
+    """
     def __init__(self, connection):
+        """Luokan konstruktori.
+
+        Args:
+            connection: Tietokantayhteyden Connection-olio
+        """
         self.connection = connection
 
     def create(self, task, user):
+        """Tallentaa tehtävän tietokantaan.
+
+        Args:
+            task: Tallennettava tehtävä Task-oliona.
+            user: Käyttäjä User-oliona.
+        """
         cursor = self.connection.cursor()
         cursor.execute("INSERT INTO tasks (topic, category, deadline, user_id) VALUES (?, ?, ?, ?)",
                        (task.topic, task.category, task.deadline, user.id))
         self.connection.commit()
 
     def find_all_by_user(self, user):
+        """Palauttaa kaikki käyttäjän tehtävät.
+
+        Args:
+            user: Käyttäjä User-oliona.
+
+        Returns:
+            Palauttaa listan Task-olioita.
+        """
         cursor = self.connection.cursor()
         cursor.execute("SELECT * FROM tasks WHERE user_id = ?", (user.id, ))
         tasks = cursor.fetchall()
@@ -21,16 +42,33 @@ class TaskRepository():
                      task["id"], task["done"]) for task in tasks]
 
     def set_done(self, task_id):
+        """Merkkaa tehtävän tehdyksi.
+
+        Args:
+            task_id: tehtävän yksilöivä id
+        """
         cursor = self.connection.cursor()
         cursor.execute("UPDATE tasks set done = 1 WHERE id = ?", (task_id, ))
         self.connection.commit()
 
     def set_not_done(self, task_id):
+        """Palauttaa tehtävän tekemättömäksi.
+
+        Args:
+            task_id: tehtävän yksilöivä id
+        """
         cursor = self.connection.cursor()
         cursor.execute("UPDATE tasks set done = 0 WHERE id = ?", (task_id, ))
         self.connection.commit()
 
     def validate_task(self, topic, category, deadline):
+        """Validoi uuden tehtävän aiheen, kategorian ja määräpäivän ennen luomista.
+
+        Args:
+            topic: Merkkijonoarvo, joka kuvaa tehtävän aihetta.
+            category: Merkkijonoarvo, joka kuvaa tehtävän kategoriaa.
+            deadline: Merkkijonoarvo, joka kuvaa tehtävän määräpäivää.
+        """
         if not topic and not category and not deadline:
             raise ValueError("Kaikki kentät ovat pakollisia")
         if not topic or len(topic.strip()) == 0: